+++ app/src/test/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionUpdateTest.kt
@@ -52,11 +52,50 @@ class TeleconsultPrescriptionUpdateTest {
         .whenEvent(DataForNextClickLoaded(
             medicalInstructions = medicalInstructions,
             medicalRegistrationId = medicalRegistrationId,
-            hasSignatureBitmap = false
+            hasSignatureBitmap = false,
+            hasMedicines = true
         ))
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(ShowSignatureRequiredError)
+        ))
+  }
+
+  @Test
+  fun `when there are no medicines added, then show medicines required error`() {
+    val medicalInstructions = "This is a medical instructions"
+    val medicalRegistrationId = "ABC12345"
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForNextClickLoaded(
+            medicalInstructions = medicalInstructions,
+            medicalRegistrationId = medicalRegistrationId,
+            hasSignatureBitmap = true,
+            hasMedicines = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowMedicinesRequiredError)
+        ))
+  }
+
+  @Test
+  fun `when there is no signature and medicines added, then show signature required & medicines required error`() {
+    val medicalInstructions = "This is a medical instructions"
+    val medicalRegistrationId = "ABC12345"
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForNextClickLoaded(
+            medicalInstructions = medicalInstructions,
+            medicalRegistrationId = medicalRegistrationId,
+            hasSignatureBitmap = false,
+            hasMedicines = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowSignatureRequiredError, ShowMedicinesRequiredError)
         ))
   }
 
@@ -71,6 +110,7 @@ class TeleconsultPrescriptionUpdateTest {
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(LoadDataForNextClick(
+                patientUuid = patientUuid,
                 teleconsultRecordId = teleconsultRecordId,
                 medicalInstructions = medicalInstructions,
                 medicalRegistrationId = medicalRegistrationId
@@ -104,7 +144,8 @@ class TeleconsultPrescriptionUpdateTest {
         .whenEvent(DataForNextClickLoaded(
             medicalInstructions = medicalInstructions,
             medicalRegistrationId = medicalRegistrationId,
-            hasSignatureBitmap = true
+            hasSignatureBitmap = true,
+            hasMedicines = true
         ))
         .then(assertThatNext(
             hasNoModel(),