+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Change `RemoteConfigService` to pull updates synchronously
 - Stop injecting screens in `TheActivity` via the static component
 - Add support for sending teleconsult request via SMS
+- Show medicines required error in `TeleconsultPrescriptionScreen`
 
 ### Fixes
 - Fix deeplink screens & warning dialogs displaying again after activity restart.